Add basic SQLite auth
