+++ taintedpaint/app/KanbanClient.tsx
@@ -1,0 +1,9 @@
+// app.page.tsx 
+
+"use client"
+
+import KanbanBoard from "@/kanban-board"
+
+export default function KanbanClientPage() {
+  return <KanbanBoard />
+}
